requirements.txt
Django~=4.0.5
django-environ
django-grappelli
django-crispy-forms
django-allauth
django-comments-xtd
django-post-office
django_mailbox
django-reset-migrations
psycopg2
django-ical


pydantic
celery~=5.2.7
uuid~=1.30
django-taggit
django-ninja
datetime~=4.5
django-model-utils~=4.2.0
django-phonenumber-field
phonenumbers
nanoid

whitenoise
wheel

#Dev packages
django-werkzeug
django-debug-toolbar~=3.5.0
# factory==1.2
pytest==7.1.2
